created CryptoMarketSummary component to clean up the CoinPage
