+++ src/components/CryptoMarketSummary/index.js
@@ -1,0 +1,1 @@
+export { default } from './CryptoMarketSummary.js'
