+++ src/components/index.js
@@ -34,3 +34,4 @@ export { default as SideArrow } from './SideArrow'
 export { default as CryptoLinks } from './CryptoLinks'
 export { default as ToolTip } from './ToolTip'
 export { default as CryptoSummary } from './CryptoSummary'
+export { default as CryptoMarketSummary } from './CryptoMarketSummary'
